Move p1_monitor coordinator to separate module
