+++ tests/components/p1_monitor/test_init.py
@@ -29,7 +29,7 @@ async def test_load_unload_config_entry(
 
 
 @patch(
-    "homeassistant.components.p1_monitor.P1Monitor._request",
+    "homeassistant.components.p1_monitor.coordinator.P1Monitor._request",
     side_effect=P1MonitorConnectionError,
 )
 async def test_config_entry_not_ready(
